first fixes

+++ sign_and_notarize.sh
@@ -11,7 +11,9 @@ RESET='\033[00m'
 ######################## Actual code, no need to edit unless you know what you're doing :) ########################
 
 echo -e "Moving the previously signed binary to trash"
-trash -F "${BINARY_SIGNED}" || { echo "${RED}Remove failure${RESET} You might want to give this terminal 'Automation' permission from System Preferences."; exit 1; }
+if [ -f "${BINARY_SIGNED}" ]; then
+	trash -F "${BINARY_SIGNED}" || { echo "${RED}Remove failure${RESET} You might want to give this terminal 'Automation' permission from System Preferences."; exit 1; }
+fi
 
 echo -e "Copying binary to work on..."
 cp -r "$BINARY_RAW" "$BINARY_SIGNED"
@@ -24,10 +26,10 @@ echo -e "Giving xr permission to everything in the binary"
 chmod -R a+xr "$BINARY_SIGNED"
 
 echo -e "Signing Dylib"
-libfile=`find "./$BINARY_SIGNED" -name '*.dylib'`
+libfile=$(find "./$BINARY_SIGNED" -name '*.dylib')
 if [ -n "$libfile" ]; then
 	while read -r libname; do
-		echo -e "${GREEN}Sign:" $libname $RESET
+		echo -e "${GREEN}Sign:" "$libname" "$RESET"
 		codesign --deep --force --verify -vvvv --timestamp --options runtime --entitlements  "$ENTITLEMENTS" --sign "$SIGNCERT" "$libname" || { echo -e "${RED}Codesign failure!${RESET}"; exit 1; }
 		
 	done <<< "$libfile"
@@ -36,10 +38,10 @@ else
 fi
 
 echo -e "Signing Bundle"
-bundlefile=`find "./$BINARY_SIGNED" -name '*.bundle'`
+bundlefile=$(find "./$BINARY_SIGNED" -name '*.bundle')
 if [ -n "$bundlefile" ]; then
 	while read -r bundlename; do
-		echo -e "${GREEN}Sign:" $bundlename $RESET
+		echo -e "${GREEN}Sign:" "$bundlename" "$RESET"
 		codesign --deep --force --verify -vvvv --timestamp --options runtime --entitlements  "$ENTITLEMENTS" --sign "$SIGNCERT" "$bundlename" || { echo -e "${RED}Codesign failure!${RESET}"; exit 1; }
 	done <<< "$bundlefile"
 else
@@ -47,10 +49,10 @@ else
 fi
 
 echo -e "Signing Framework"
-frameworkfile=`find "./$BINARY_SIGNED" -name '*.framework'`
+frameworkfile=$(find "./$BINARY_SIGNED" -name '*.framework')
 if [ -n "$frameworkfile" ]; then
 	while read -r frameworkname; do
-		echo -e "${GREEN}Sign:" $frameworkname $RESET
+		echo -e "${GREEN}Sign:" "$frameworkname" "$RESET"
 		codesign --deep --force --verify -vvvv --timestamp --options runtime --entitlements  "$ENTITLEMENTS" --sign "$SIGNCERT" "$frameworkname" || { echo -e "${RED}Codesign failure!${RESET}"; exit 1; }
 	done <<< "$frameworkfile"
 else
@@ -58,7 +60,7 @@ else
 fi
 
 echo -e "Signing the entire app..."
-codesign --deep --force --verify --verbose --timestamp --options runtime --entitlements  "$ENTITLEMENTS" --sign "$SIGNCERT" "$BINARY_SIGNED" -i $BUNDLE_ID
+codesign --deep --force --verify --verbose --timestamp --options runtime --entitlements  "$ENTITLEMENTS" --sign "$SIGNCERT" "$BINARY_SIGNED" -i "$BUNDLE_ID"
 
 echo -e "Verifications...."
 codesign --verify --deep -vvvv --strict "$BINARY_SIGNED" || { echo -e "${RED}Codesigning verification failure!${RESET}"; exit 1; }
@@ -69,9 +71,8 @@ echo -e "Zipping..."
 /usr/bin/ditto -c -k --sequesterRsrc --keepParent "$BINARY_SIGNED" "$ZIP_NAME" || { echo -e "${RED}Zipping failed!${RESET}"; exit 1; }
 
 echo -e "Sending for notarization..."
-xcrun altool --notarize-app --asc-provider $PROVIDERSHORTNAME  --primary-bundle-id "$BUNDLE_ID" --username "$APPLE_ID" --password "@keychain:${KEYCHAIN_ITEM_ID}" --file "$ZIP_NAME"
 
-if [ $? -eq 0 ]; then
+if xcrun altool --notarize-app --asc-provider "$PROVIDERSHORTNAME"  --primary-bundle-id "$BUNDLE_ID" --username "$APPLE_ID" --password "@keychain:${KEYCHAIN_ITEM_ID}" --file "$ZIP_NAME"; then
 	echo -e "${GREEN}Success! Your app is being notarized.${RESET}"
 	echo "Go grab some coffee and check back later with command xcrun altool --username \"${APPLE_ID}\" --password \"@keychain:${KEYCHAIN_ITEM_ID}\" --notarization-info <replace this with the request uuid that appears above>"
 	exit 0
